# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Claude Code Zen API
  description: |
    Comprehensive API documentation for Claude Code Zen - AI-powered development coordination platform.
    
    This API provides endpoints for:
    - System health and status monitoring
    - Swarm management and coordination
    - Task management and orchestration
    - Document and file operations
    - LLM statistics and analytics
    - Matron advisory system
    - Visionary roadmap planning
    - Workspace management
    - Real-time WebSocket communications
    
    The API follows RESTful principles and returns JSON responses.
  version: 2.0.0
  contact:
    name: Claude Code Zen Team
    url: https://github.com/zen-neural/claude-code-zen
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://fra-d1.in.centralcloud.net/api
    description: Production server
paths:
...
  /status:
    get:
      summary: System Status
      description: Get comprehensive system status including performance metrics
      operationId: getSystemStatus
      tags:
        - System
      responses:
        '200':
          description: System status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
        '500':
          description: Failed to get system status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Swarm Management
...
  /execute:
    post:
      summary: Execute Command
      description: Execute a system command or script
      operationId: executeCommand
      tags:
        - Command Execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteCommandRequest'
            example:
              command: "npm"
              args: ["test", "--coverage"]
              workingDirectory: "/project/root"
      responses:
        '200':
          description: Command executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResult'
        '400':
          description: Invalid command
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Command execution failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Settings Management
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code (optional)
          example: "NOT_FOUND"
        details:
          type: object
          description: Additional error details (optional)

    # System Status
    SystemStatus:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          example: "healthy"
        timestamp:
          type: string
          format: date-time
        uptime:
          type: number
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceStatus'
    PerformanceMetrics:
      type: object
      properties:
        cpu:
          type: number
          format: float
          description: CPU usage percentage
          example: 15.5
        memory:
          type: number
          format: float
          description: Memory usage percentage
          example: 45.2
        disk:
          type: number
          format: float
          description: Disk usage percentage
          example: 60.1
        responseTime:
          type: number
          description: Average response time in milliseconds
          example: 125
    ServiceStatus:
      type: object
      properties:
        name:
          type: string
          example: "swarm-coordinator"
        status:
          type: string
          enum: [running, stopped, error]
          example: "running"
        lastCheck:
          type: string
          format: date-time
        details:
          type: object

    # Swarm Management
...
    ExecuteCommandRequest:
      type: object
      required:
        - command
      properties:
        command:
          type: string
          example: "npm"
        args:
          type: array
          items:
            type: string
          example: ["test", "--coverage"]
        workingDirectory:
          type: string
          example: "/project/root"
        environment:
          type: object
          additionalProperties:
            type: string
          example:
            NODE_ENV: "test"
        timeout:
          type: integer
          description: Command timeout in seconds
          default: 30
          example: 60
    CommandResult:
      type: object
      properties:
        command:
          type: string
          example: "npm"
        args:
          type: array
          items:
            type: string
          example: ["test", "--coverage"]
        exitCode:
          type: integer
          example: 0
        stdout:
          type: string
          example: "Test output..."
        stderr:
          type: string
          example: ""
        success:
          type: boolean
          example: true
        duration:
          type: number
          description: Command execution time in milliseconds
          example: 5420

    # Settings
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (if implemented)
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...